profil et historique

+++ src/main/java/com/projet_voiture/projet_voiture/modele/NombreVoitureVenduParMois.java
@@ -11,6 +11,7 @@ public class NombreVoitureVenduParMois {
     
     private int anneeValidation;
     private int moisValidation;
-    private int nombreValidations;
+    private long nombreValidations;
+    
 
 }
